# public/api/openapi.yaml
openapi: 3.0.0
info:
  title: Book Review API
  description: API for managing books, reviews, and user interactions
  version: 1.0.0
  contact:
    email: y.kabir1@edu.salford.ac.uk
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        profilePicture:
          type: string
          nullable: true
          example: "/uploads/profile_pictures/john-doe.jpg"
        roles:
          type: array
          items:
            type: string
          example: ["ROLE_USER"]
        banned:
          type: boolean
          example: false
        readingList:
          type: array
          items:
            $ref: '#/components/schemas/Book'
          example: []
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          example: []
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        googleBookId:
          type: string
          example: "zyTCAlFPjgYC"
        title:
          type: string
          example: "The Great Gatsby"
        authors:
          type: array
          items:
            type: string
          example: ["F. Scott Fitzgerald"]
        thumbnail:
          type: string
          nullable: true
          example: "https://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1&source=gbs_api"
        description:
          type: string
          nullable: true
          example: "The story of the mysteriously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan."
        pageCount:
          type: integer
          nullable: true
          example: 180
        categories:
          type: array
          items:
            type: string
          example: ["Fiction", "Classics", "Literary"]
        averageRating:
          type: number
          format: float
          nullable: true
          example: 4.5
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
          example: []
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        content:
          type: string
          example: "A masterpiece of American literature. The prose is beautiful and the story is timeless."
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2023-05-15T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-05-15T14:30:00Z"
        user:
          $ref: '#/components/schemas/User'
          example: {"id": 1, "name": "John Doe"}
        book:
          $ref: '#/components/schemas/Book'
          example: {"id": 1, "title": "The Great Gatsby"}
        likesCount:
          type: integer
          example: 12
        likedByCurrentUser:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example: {"email": ["This value is not a valid email address."]}
        code:
          type: integer
          example: 400
      required:
        - success
        - message
        - code
  responses:
    UnauthorizedError:
      description: Unauthorized - User is not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            code: 401
            errors: {}
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Resource not found"
            code: 404
            errors: {}
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation failed"
            code: 400
            errors:
              email: ["This value is not a valid email address."]
              password: ["This value is too short. It should have 8 characters or more."]
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Resource already exists"
            code: 409
            errors: {}
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "An unexpected error occurred"
            code: 500
            errors: {}
paths:
...
  /books/search:
    get:
      summary: Search books
      tags:
        - Books
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
          example: "The Great Gatsby"
      responses:
        '200':
          description: Books found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
              example:
                - id: 1
                  googleBookId: "zyTCAlFPjgYC"
                  title: "The Great Gatsby"
                  authors: ["F. Scott Fitzgerald"]
                  thumbnail: "https://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1&source=gbs_api"
                  description: "The story of the mysteriously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan."
                  pageCount: 180
                  categories: ["Fiction", "Classics", "Literary"]
                  averageRating: 4.5
                  reviews: []
                - id: 2
                  googleBookId: "1q2w3e4r5t6y"
                  title: "Gatsby: A Novel"
                  authors: ["John Smith"]
                  thumbnail: "https://books.google.com/books/content?id=1q2w3e4r5t6y&printsec=frontcover&img=1&zoom=1&source=gbs_api"
                  description: "A modern retelling of the classic story."
                  pageCount: 250
                  categories: ["Fiction", "Modern"]
                  averageRating: 3.8
                  reviews: []
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /books/{googleBookId}:
    get:
      summary: Get book details
      tags:
        - Books
      parameters:
        - name: googleBookId
          in: path
          required: true
          schema:
            type: string
          example: "zyTCAlFPjgYC"
        - name: review_page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Review page number
          example: 1
        - name: review_limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
          description: Reviews per page
          example: 10
        - name: review_sort
          in: query
          schema:
            type: string
            enum: [createdAt, rating, likesCount]
          description: Review sort field
          example: "createdAt"
        - name: review_direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
          description: Review sort direction
          example: "desc"
      responses:
        '200':
          description: Book details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
                  googleBookId:
                    type: string
                    example: "zyTCAlFPjgYC"
                  title:
                    type: string
                    example: "The Great Gatsby"
                  authors:
                    type: array
                    items:
                      type: string
                    example: ["F. Scott Fitzgerald"]
                  thumbnail:
                    type: string
                    nullable: true
                    example: "https://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1&source=gbs_api"
                  description:
                    type: string
                    nullable: true
                    example: "The story of the mysteriously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan."
                  pageCount:
                    type: integer
                    nullable: true
                    example: 180
                  categories:
                    type: array
                    items:
                      type: string
                    example: ["Fiction", "Classics", "Literary"]
                  averageRating:
                    type: number
                    format: float
                    nullable: true
                    example: 4.5
                  reviews:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                            example: 25
                          page:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 10
                          total_pages:
                            type: integer
                            example: 3
                          has_next:
                            type: boolean
                            example: true
                          has_prev:
                            type: boolean
                            example: false
              example:
                id: 1
                googleBookId: "zyTCAlFPjgYC"
                title: "The Great Gatsby"
                authors: ["F. Scott Fitzgerald"]
                thumbnail: "https://books.google.com/books/content?id=zyTCAlFPjgYC&printsec=frontcover&img=1&zoom=1&source=gbs_api"
                description: "The story of the mysteriously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan."
                pageCount: 180
                categories: ["Fiction", "Classics", "Literary"]
                averageRating: 4.5
                reviews:
                  data: [
                    {
                      id: 1,
                      content: "A masterpiece of American literature. The prose is beautiful and the story is timeless.",
                      rating: 5,
                      createdAt: "2023-05-15T14:30:00Z",
                      updatedAt: "2023-05-15T14:30:00Z",
                      user: {"id": 1, "name": "John Doe"},
                      likesCount: 12,
                      likedByCurrentUser: true
                    }
                  ]
                  pagination:
                    total: 25
                    page: 1
                    limit: 10
                    total_pages: 3
                    has_next: true
                    has_prev: false
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /books/open-library/{key}:
    post:
      tags:
        - Books
      summary: Add a book from Open Library
      description: Fetches book data from Open Library API and adds it to the database
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Open Library book key (e.g., W123456)
      responses:
        '201':
          description: Book successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Book added successfully
                  data:
                    type: object
                    properties:
                      book:
                        $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/ServerError'
...